G25Tools

Introduction

G25Tools is a library for the Julia programming language to work with G25 coordinates.

G25 coordinates were introduced by the blogger Davidski to compare genetic samples and populations.

First steps

Download the main file of ancient samples from ExploreYourDNA.

I also highly recommend to get your G25 coordinates from G25 Requests. If you already know your K36 coordinates you can also begin with simulated G25 coordinates from ExploreYourDNA. Note that this step is not necessary to get the examples running.

Start working with the examples.

Functions

G25Tools.Distances — Type

Distances

Holds the distances from multiple source samples to a single
target sample. Each source sample consists of a name and it's
G25 distance to the target sample.

G25Tools.Population — Type

Population

Represents a population. Each population consists of a name,
for example "Viking", and a list of samples in G25 format.

G25Tools._cluster — Method

_cluster(index::Integer, indices::Integer[], distancematrix, distance)

Return all indices that cluster with the index index.

G25Tools._fix_lat_long — Method

_fix_lat_long(rawdata::AbstractDataFrame)

Make sure that the Lat and Long columns in the loaded data contain only Float64 or missing entries.

The original CSV file with ancient samples sometimes contains multiple GPS entries for a single? (fused samples?) sample. This results in invalid entries when the file is loaded by the CSV package.

G25Tools._getmarkers — Method

_getmarkers(n::Integer; smile = false)

Create a set of n markers for a plot. If smile == true the last marker is a smiley.

Return a named tuple (markers, colors) that contains an array of markers and an array of colors for the plot.

G25Tools._select — Method

_select(samples::DataFrame, distance, neighbors)

Select those samples that have at least the given number of neighbors within the given distance. The distancematrix consists of sample indices and their distances to each other.

Return an array sample indices that satisfy the given conditions.

G25Tools.average — Method

average(samples::DataFrame, name = "Average")

Calculate the mean average of a DataFrame of samples in G25 format.

Return the average as a vector of G25 coordinates where the first entry is the given name, usually the name of the given population.

G25Tools.clusters — Method

clusters(samples::DataFrame; distance = 0.1, neighbors = 1)

Find clusters in a set of samples that satisfy the given conditions. Within each cluster the nearest neighbor of a sample must be within distance and each sample must have a minumum number of neighbors that is given by neighbors.

Return an array of Populations where each Population represents one cluster. If no cluster could be found an empty array is returned.

XXX Currently each sample may appear in only one cluster, which can lead to strange results. Also samples at the fringe of a cluster are sometimes cut off if they do not have enough neighbors. Performance is rather poor for large sample sizes. About 6 seconds for 500 samples on my rather old computer. 24 seconds for 600 samples.

G25Tools.distance — Method

distance(source::DataFrameRow, target::DataFrameRow)

Return the vector distance between two samples with G25 coordinates. Both source and target must be in the form: ["Name", PC1,...PC25] where Name is a String and PC1 to PC25 are the G25 coordinates of the sample. Additional fields are allowed because only [2:26] are used to calculate the distance.

G25Tools.distance — Method

distance(source::Vector, target::Vector)

Return the vector distance between two samples with G25 coordinates. Both source and target must be in the form: ["Name", PC1,...PC25] where Name is a String and PC1 to PC25 are the G25 coordinates of the sample. Additional fields are allowed because only [2:26] are used to calculate the distance.

G25Tools.distances — Method

distances(source::DataFrame, target::DataFrameRow)

Calculate the G25 vector distances from a list of samples (source) to a single target.

Return a DataFrame containing the distances to the target. Each row consists of the source name and the distance to the target. The table is ordered, smallest distances come first.

G25Tools.distances — Method

distances(source::DataFrame, target::DataFrame)

Calculate the G25 vector distances from a list of samples (source) to multiple target samples.

The source and the target contain a DataFrame in G25 format.

Return a vector of Distances where each entry contains the Distances of a single target sample to all source samples.

G25Tools.distances — Method

distances(source::DataFrame, target::Vector)

Calculate the distances between samples in a DataFrame and a single target sample.

G25Tools.extractG25 — Method

extractG25(filename::AbstractString, to::AbstractString; cols = String[])

Extract G25 data from the main file and save it locally to disk. The main file can be found at: https://www.exploreyourdna.com/ancient-samples.aspx

The G25 data is a simplified version of the main file which ir rather large to work with. Also some population data is only available in the simplified format.

The parameter cols can be used to add extra columns from the main file, for example it may be useful to add GPS coordinates. Just add the column titles as a String array.

G25Tools.extractG25 — Method

extractG25(filename::AbstractString; cols = String[])

Extract G25 data from the main file of ancient samples. The main file can be found at: https://www.exploreyourdna.com/ancient-samples.aspx

Return a DataFrame in a simplified format that represents samples as Names and their G25 coordinates.

The parameter cols can be used to add extra columns from the main file, for example it may be useful to add GPS coordinates. Just add the column titles as a String array.

G25Tools.extractG25 — Method

function extractG25(from::AbstractDataFrame; cols = String[])

Extract G25 data from a DataFrame that is in the form of the main file of ancient samples. The main file can be found at: https://www.exploreyourdna.com/ancient-samples.aspx

Return a DataFrame in a simplified format that represents samples as Names and their G25 coordinates.

The parameter cols can be used to add extra columns from the main file, for example it may be useful to add GPS coordinates. Just add the column titles as a String array like cols = ["Lat.", "Long."].

G25Tools.getyear — Method

getyear(samplename)

Gets the year from a sample name in Davidskis format. Davidski encodes the year in the sample name with a pattern like BC_50 or AD_450.

Returns the year or missing if the year is not encoded.

G25Tools.medianavg — Method

median(samples::DataFrame; name = "Median")

Calculate the median average of a DataFrame of samples in G25 format.

Return the median average as a vector of G25 coordinates where the first entry is the given name, usually the name of the given population.

This method may be better suited than the mean aveerage for small sample sizes that include outliers.

G25Tools.picksamples — Method

function picksamples(samples::AbstractDataFrame; populations = String[], timeperiod = [-Inf, Inf])

Pick a table of samples belonging to certain populations and/or a time persiod. Example: picksamples(samples, populations=["Germany", "England"], timeperiod=[0, 500])

For more advanced filtering options use DataFrames' subset() or filter() methods.

G25Tools.plotpca — Method

plotpca(populations::Array{Population}; title = "PCA Plot", smile = false)

Plot all populations. The names of the populations are used in the plot legend. If smile == true the last marker is a smiley.

This method generates a label for each population.

G25Tools.plotpca — Method

plotpca(samples::DataFrame)

Plot all samples in a given DataFrame. The DataFrame must be in the G25 format: Name followed by PC1,...PC25. If smile == true the last marker is a smiley.

This method generates a label for each sample.

G25Tools.readG25 — Method

readG25(filename)

Read a DataFrame containing a list of samples with G25 coordinates. Each row must begin with the sample name followed by their G25 coordinates. Example: Name,PC1,PC2,...PC25

G25Tools.topmatches — Method

topmatches(sourcesamples, targetsamples;
           sourcetimeperiod = [-Inf, Inf], targettimeperiod = [-Inf, Inf],
           maxdistance = 0.1, targetfraction = 0.68)

Find sourcesamples that had a big impact on the targetsamples. The intention is to find the target's most important ancestral population.

The method looks for matches within the given G25 distance (maxdistance) to the target samples. It selects those matches that match the specified number of targetsamples (targetfraction) or more.

G25Tools.writedistances — Method

writedistances(directory, filename, distances)

Write a DataFrame in G25 format to a specific directory. This is a convenience function that is most useful when writing multile files to a single directory.

Index

G25Tools.Distances

G25Tools.Population

G25Tools._cluster

G25Tools._fix_lat_long

G25Tools._getmarkers

G25Tools._select

G25Tools.average

G25Tools.clusters

G25Tools.distance

G25Tools.distance

G25Tools.distances

G25Tools.distances

G25Tools.distances

G25Tools.extractG25

G25Tools.extractG25

G25Tools.extractG25

G25Tools.getyear

G25Tools.medianavg

G25Tools.picksamples

G25Tools.plotpca

G25Tools.plotpca

G25Tools.readG25

G25Tools.topmatches

G25Tools.writedistances